backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
requests==2.31.0
reportlab==4.0.4
python-dotenv==1.0.0
marshmallow==3.20.1
werkzeug==2.3.7